Get the tab size from our options instead of from the text buffer directly
In 2f2066708 I changed CodeModel to not open an editor when somebody
called GetStartPoint(), but it turns out that some paths of
GetStartPoint() internally called GetTabSize which assumed the text
was backed by a text buffer. This was the original reason the
EnsureEditor call was added, but I incorrectly determined that the
it was no longer needed. Now we just pass our OptionSet from the
document down and ask that, which will work in all cases.