Introduce Exec-Command
There is no good method today for executing windows commands with
dynamically built argument lists in Powershell.  This change attempts
to address that by introducing Exec-Command which uses the .NET
Process API directly.  This makes it straight forward to build windows
commands that rely on windows style parsing and output