Use Interlocked exchange on some MetadataReferences to guarantee reference equality (#18782)
It seems like the reference set for CoreCLR tests is sometimes different, resulting
in CI failures. This is likely due to the lazy initialization of static MetadataReferences
in TestBase without thread synchronization. This change adds Interlocked calls where
necessary and adds a comment warning about when more Interlocked calls may be
needed.

This is a test-only change.