Acquire the user's selection up-front if we're on the UI thread for light-bulbs.
There are circumstances where the editor will call into us on hte UI thread and will then
block for results.  Right now this can deadlock as we may have kicked off work to the BG
that then awaits getting the selection back from the UI thread.  To avoid that, if we're on
the UI thread when we start, just go and get the selection and pass it along.