Consider name when matching hierarchy items to projects
Our support for showing nodes for analyzers and diagnostics in the Solution Explorer depends on us being able to map from an `IVsHierarchy` and item ID representing a project node to the corresponding Roslyn project in the workspace. In Dev14 this was relatively easy, as we could simply check each `IVsHierarchy` held by a Roslyn project for reference equality with the `IVsHierarchy` for the project in Solution Explorer.

This approach doesn't work in Dev15 for CPS-based projects. Because of multi-targeting there is a one-to-many mapping from a project in Solution Explorer to projects in the Roslyn workspace; each project in the Roslyn workspace holds on to a different, target-framework-specific `IVsHierarchy` that comes from the project system and is never seen by Solution Explorer. To handle this I switched to using the "canonical name" and target framework to match things up.

It turns out this breaks the legacy project system when two projects are in the same folder--they both report the _directory path_ as their canonical name, rather than their full path. The code to map from Solution Explorer nodes to Roslyn projects thus ends up finding two matching projects where it expected only one. The end result is that the project loads, but when you try to expand it in Solution Explorer you get no children.

To fix this we need to check both the canonical name, and the regular name. Even if two projects share the same directory, their names will necessarily be different.