Fix the TestVsiNetCore runs
This moves the logic for testVsiNetCore into the BuildAndTest.proj file.
This matches the logic behavior for all of our other test runs.  It also
avoids us further complicating the necessary command line for calling
the BuildAndTest.proj file.

Eventually though want to move all of the test logic into a Powershell
file.  The MSBuild file at this point is approaching the complexity of a
full script but lacking the flexibility.  Should just convert it to a
script and move on.