Parsing of readonly structs.
`readonly` is already a legal modifier in the grammar.
The only change here is really to not reject `readonly` on structs in semantics.

Also, produce lang version errors now if `readonly` is used on a struct declaration with old lang versions.

NOTE: We are not adding any meaning to readonly structs in this change.