Process documents for open after a moniker change
With the legacy project system, it just so happened that a file rename
would be updated in the RDT prior to the project system removing the
file under the old name and adding it under the new name. The new
project system doesn't guarantee such an ordering (nor does it have any
reason to -- either ordering is perfectly reasonable), so let's handle
that too.

Fixes dotnet/roslyn#19389.