Add extra logging to BuildBoss
See the following PR for motivation here

> https://github.com/dotnet/roslyn/pull/19787

Essentially the windows build correctness leg is failing due to an error
in BuildBoss.  There is no output associated with this error, just a
failing return code.

Looking at the BuildBoss code the only plausible explanation for this
behavior is an unhandled exception.  Added code to catch this case and
print out a reasonable error when it occurs.