Ensure that IOperations nodes within a C# local function are part of the operation tree
Analysis of any code within a local function with IOperation analyzers will produce false or missing diagnostics, as the operation tree doesn't contain any operation nodes within the BoundLocalFunctionStatement. This can be a dogfood blocker for IOperation analyzers on real world code - we had few false reports due this on the Roslyn repo as we use this new language feature, and corresponding rules had to be disabled in the repo.