Do not call GetOpenDocumentInCurrentContextWithChanges(...) in AbstractMemberInsertingCompletionProvider
I noticed this while working through an issue using the CompletionService in OmniSharp.

`GetOpenDocumentInCurrentContextWithChanges(...)` allows the caller to retrieve a Document from a SourceText. It achieves this be calling `Workspace.TryGetWorkspace(...)` with the SourceText's SourceTextContainer. However, in the case of a standard SourceText, this is just a StaticContainer that changes every time `SourceText.WithChanges(...)` is called. So, if the SourceText is updated for a particular Document in an AdhocWorkspace, calling `GetOpenDocumentInCurrentContextWithChanges(...)` for the new SourceText will fail because its container is different than the one that was initially registered with the Workspace.