Removed TestUtilities.Desktop deps in solutions
In theory this should only insert a build order dependency between these
projects.  Forcing TestUtilities.Desktop to build before the parent
project.

In the new SDK this is treated as an actual reference, not a build order
dependency.  This means that it eventually sees a NetStandard project
referencing a Desktop one and errors out.