Move language settings after SDK import
The new SDK unconditionally sets a number of MSBulid properties
including many that we depend on in Roslyn.  This is a break from the
old MSBuild behavior which respected initial values.

https://github.com/dotnet/sdk/issues/752

The behavior in 2.0 is changing to go back to the old model of
respecting initial values.  Roslyn though continues to often use the SDK
installed with Visual Studio.  That can be a variety of values from 2.0
prebuilds, to various drops of 1.0.

As a result the success / failure of our build can depend a lot on which
version of VS is installed and by consequence which new SDK comes with
it.  Subtle differences in the SDK can change the values we end up
compiling with in our project.  We've been guarding against this
by directly re-setting values we knew to be overidden in some version
of the SDK.

Now though I'm worried this isn't an adequete solution for the following
reasons.

- Just this week I got new warnings in the build by using a new VS
- The Microbuild lab is aggressively grabbing new builds of VS which
means the VS SDK is also potentially changing.

After some thought I decided to simply move all of our langauge specific
settings after we import the core targets.  Given that we
unconditionally set the values it becomes irrelevant what the SDK does
or does not respect.