Suppress VB 40057 at a global level
The warning B40057 used to be suppressed at a global level in PCL /
Desktop projects. The SDK does not suppress this warning and hence
converting to the new SDK caused this to show up.

Previously a number of developers fixed this problem individually. The
correct solution is to suppress it globally as was done previously.