Fixing a race in PEMethodSymbol
Fixes:#20496

The race seems to have more effect when running on CoreClr/Linux and in particular affects CoreFxLab repo where it causes build breaks.

There is a test that this race causes to fail - TestOverridingBaseConstructorStruct, but it happens very rarily.

The issue seems to be in computation of the use site diagnostics in PEMethodSymbol
One of possible race scenarios:

thread #1 sees that IsUseSiteDiagnosticPopulated is not set, goes to compute and publish diagnostics
thread #2 sees that IsUseSiteDiagnosticPopulated is not set, goes to compute and publish diagnostics
thread #1 finds no issues and sets IsUseSiteDiagnosticPopulated
thread #3 does something else and indirectly creates UncommonFields with a sentinel value for the use site diagnostics
thread #2 checks again IsUseSiteDiagnosticPopulated and since it is set (by thread #1) happily returns the sentinel as the actual value!!!

While we must fix the race ASAP in the readonly-ref branch, the rest of the PEMethodSymbol must be reviewed/simplified.
It seems to reuse this same pattern that mixes sentinel/null/flag as indication of the current state thus introducing a lot of states and it is hard to prove that incoherent states are unreachable.