Go through the unconfigured project to get the configured project
We're trying to go from an `IVsBrowseObjectContext` to an `IPackageReferencesService`.
We were doing this through `IVsBrowseObjectContext.ConfiguredProject`, but it turns out
that this property always returns null on the particular implementation of
`IVsBrowseObjectContext` we're working with.

Instead, we need to go through the `UnconfiguredProject` and ask it for the "suggested"
`ConfiguredProject`.

I've also switched from `GetAwaiter().GetResult()` to properly async/await code inside
an `ExecuteSynchronously` call, mostly for the purposes of readability.