Don't try to take a screenshot if there is no screen
There are cases where we could inadvertently end up trying to run
integration tests but there is no screen. For example, when the user
session is not, for whatever reason, an interactive session. If this
happens and we invoke the `ScreenshotService` we'll try to create a
zero width, zero height `Bitmap` and end up throwing an exception.

If there is no screen we expect lots of things to fail badly, and
there is no point in trying to capture a screenshot and adding another
failure on top of everything else. Here we check if there is actually
a screen to capture, and bail out early if not.