Avoid recursive exceptions when logging exceptions
In our integration tests we hook up a handler to `AppDomain.CurrentDomain.FirstChanceException`
to allow us to capture screenshots when something goes wrong in a test. This handler must not
throw an exception itself, or it will end up being called recursively. This could well overflow
the stack.

When I wrote this code I attempted to protect against this with a try/catch block. However, the
`FirstChanceException` handler is invoked *before* any exception handling, so the try/catch does
nothing.

Instead, we now use a thread-static field to record whether or not we're already in the handler.
If we are, we simply bail out without doing anything.