Cleanup the test-determinism script
The determinism test script was authored when our build still dumped everything into
a common directory. That meant the process of finding the DLLs / EXEs to check for
deterministic output was rather complex because we had to sort out the DLLS / EXES
we weren't building.

Now that our build is more structured it's simple to find the content we are directly
building: just search the Obj directory.