Ensure we disconnect any projection buffers we create for views
When we create a projection buffer, that buffer is strongly held by
the underlying buffer, which means we have a memory leak unless we
disconnect it by unmapping all spans. This also causes problems because
bugs in the editor around projections can then cause crashes when we
didn't actually need the buffers anymore.