Fix PIA embedding for Nuget.VisualStudio.
The MarkNuGetVisualStudioPackageWithEmbedInteropType target was depending on a ResolveNuGetPackageAssets target that no longer runs as part of Roslyn builds since Roslyn has been switched over to use the new SDK. Consequently, the Nuget.VisualStudio PIA was not getting embedded in MicroSoft.VisualStudio.LanguageServices.dll. Switching the dependency to the ResolvePackageDependenciesForBuild target should fix this issue.